Wait child process terminate after sending killsig
A custom termination signal can be configured with --killsig argument to
be sent to the child process, which can handle it for proper
termination. However, right after such signal is sent, child is marked
to be deleted by `DeleteConnection()` through ~processClass(), which
will SIGKILL the process. This double-killing can take milliseconds and
gives no room for a child handler to be completely executed, besides
leading to multiple SIGKILL when default --killsig=9 is used.

Wait for a SIGCHLD up to half a second before forcibly terminating it. A
timeout of 0.5s has been chosen to keep consistent with the amount of
delay already perceived for other child interactions, including normal
termination and restart.